+++ resources/customized_payloads/Rafale_M.lua
@@ -1,0 +1,209 @@
+local unitPayloads = {
+	["name"] = "Rafale_M",
+	["payloads"] = {
+		[1] = {
+			["name"] = "CAP",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "LAU-115_2*LAU-127_AIM-120C",
+					["num"] = 2,
+				},
+				[4] = {
+					["CLSID"] = "LAU-115_2*LAU-127_AIM-120C",
+					["num"] = 9,
+				},
+				[5] = {
+					["CLSID"] = "{40EF17B7-F508-45de-8566-6FFECC0C1AB8}",
+					["num"] = 8,
+				},
+				[6] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 5,
+				},
+				[8] = {
+					["CLSID"] = "{40EF17B7-F508-45de-8566-6FFECC0C1AB8}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 11,
+			},
+		},
+		[2] = {
+			["name"] = "CAS",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "LAU3_HE5",
+					["num"] = 2,
+				},
+				[4] = {
+					["CLSID"] = "LAU3_HE5",
+					["num"] = 9,
+				},
+				[5] = {
+					["CLSID"] = "LAU3_WP156",
+					["num"] = 8,
+				},
+				[6] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 5,
+				},
+				[8] = {
+					["CLSID"] = "LAU3_WP156",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 11,
+			},
+		},
+		[3] = {
+			["name"] = "STRIKE",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "{AB8B8299-F1CC-4359-89B5-2172E0CF4A5A}",
+					["num"] = 2,
+				},
+				[4] = {
+					["CLSID"] = "{AB8B8299-F1CC-4359-89B5-2172E0CF4A5A}",
+					["num"] = 9,
+				},
+				[5] = {
+					["CLSID"] = "{AB8B8299-F1CC-4359-89B5-2172E0CF4A5A}",
+					["num"] = 8,
+				},
+				[6] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 5,
+				},
+				[8] = {
+					["CLSID"] = "{AB8B8299-F1CC-4359-89B5-2172E0CF4A5A}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 11,
+			},
+		},
+		[4] = {
+			["name"] = "SEAD",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "{D5D51E24-348C-4702-96AF-97A714E72697}",
+					["num"] = 2,
+				},
+				[4] = {
+					["CLSID"] = "{D5D51E24-348C-4702-96AF-97A714E72697}",
+					["num"] = 9,
+				},
+				[5] = {
+					["CLSID"] = "{D5D51E24-348C-4702-96AF-97A714E72697}",
+					["num"] = 8,
+				},
+				[6] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 5,
+				},
+				[8] = {
+					["CLSID"] = "{D5D51E24-348C-4702-96AF-97A714E72697}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 11,
+			},
+		},
+		[5] = {
+			["name"] = "ANTISHIP",
+			["pylons"] = {
+				[1] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 10,
+				},
+				[2] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 1,
+				},
+				[3] = {
+					["CLSID"] = "{18617C93-78E7-4359-A8CE-D754103EDF63}",
+					["num"] = 2,
+				},
+				[4] = {
+					["CLSID"] = "{18617C93-78E7-4359-A8CE-D754103EDF63}",
+					["num"] = 9,
+				},
+				[5] = {
+					["CLSID"] = "{18617C93-78E7-4359-A8CE-D754103EDF63}",
+					["num"] = 8,
+				},
+				[6] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 6,
+				},
+				[7] = {
+					["CLSID"] = "{0DA03783-61E4-40B2-8FAE-6AEE0A5C5AAE}",
+					["num"] = 5,
+				},
+				[8] = {
+					["CLSID"] = "{18617C93-78E7-4359-A8CE-D754103EDF63}",
+					["num"] = 3,
+				},
+			},
+			["tasks"] = {
+				[1] = 11,
+			},
+		},
+	},
+	["tasks"] = {
+	},
+	["unitType"] = "Rafale_M",
+}
+return unitPayloads